requirements.txt
# EchoForge Requirements
# Core web framework and server
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart

# Database (using regular sqlite3 for now, fallback from sqlcipher3)
# pysqlcipher3 (if encryption needed - install manually)

# AI and ML
ollama>=0.1.0
openai-whisper>=20231117
torch>=2.0.0
numpy

# Data processing and utilities  
pyyaml>=6.0
networkx>=3.0
requests>=2.28.0
pydantic>=2.0.0

# Web and networking
websockets>=10.0
httpx>=0.24.0

# Optional dependencies for enhanced functionality
# Uncomment if needed:
# scikit-learn  # for advanced analytics
# matplotlib    # for plotting
# pandas        # for data manipulation
# plotly        # for interactive charts
 
# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
httpx==0.25.2
python-multipart==0.0.6

# Database dependencies
sqlite3
pathlib

# AI/ML dependencies (optional but recommended)
# openai-whisper==20231117  # Uncomment for voice features
# torch>=2.0.0  # Required for Whisper

# Data processing
pydantic==2.5.0
pyyaml==6.0.1
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
watchfiles==0.21.0

# Optional: Enhanced features
# networkx==3.2  # For resonance mapping analysis
# pandas>=2.0.0  # For data analysis
# numpy>=1.24.0  # For numerical operations

# Development dependencies (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
